more work for B

+++ OperationControllers/ClusterComplex.cpp
@@ -251,7 +251,7 @@ TimeSeries *meanVector(std::unordered_map<std::string, TimeSeries*> &Cluster){
 
     std::vector<__TIMESERIES_X_TYPE> *meanVec = new std::vector<__TIMESERIES_X_TYPE>;
     for(int i=0; i<tempVecSize; i++)
-        meanVec->push_back((int)tempVec[i]);
+        meanVec->push_back(tempVec[i]);
     std::string *no_s = nullptr;
     TimeSeries *meanP = new TimeSeries(meanVec, no_s);
